fix(report): resolve duplicate SARIF declaration split
